fix: simple experiment_registry
* Just store all experiments in a Dict for now
* Also factors out the Phoenix tracing setup into a separate utility not
  tied to agent creation
