+++ warehouse/oso_agent/oso_agent/util/tracing.py
@@ -1,7 +1,6 @@
 import logging
-import typing as t
+from typing import Optional
 
-from llama_index.core.agent.workflow.base_agent import BaseWorkflowAgent
 from openinference.instrumentation.llama_index import LlamaIndexInstrumentor
 from opentelemetry.exporter.otlp.proto.http.trace_exporter import (
     OTLPSpanExporter as HTTPSpanExporter,
@@ -10,18 +9,11 @@ from opentelemetry.sdk import trace as trace_sdk
 from opentelemetry.sdk.trace.export import SimpleSpanProcessor
 from phoenix.otel import register
 
-from ..util.config import AgentConfig
-from ..util.errors import AgentConfigError, AgentMissingError
-from .react_agent import create_react_agent
-from .sql_agent import create_sql_agent
+from .config import AgentConfig
 
-# Setup logging
 logger = logging.getLogger(__name__)
 
-# Type alias for a dictionary of agents
-AgentDict = t.Dict[str, BaseWorkflowAgent]
-
-def _setup_telemetry(config: AgentConfig) -> t.Optional[trace_sdk.TracerProvider]:
+def setup_telemetry(config: AgentConfig) -> Optional[trace_sdk.TracerProvider]:
     """Setup OpenTelemetry tracing if enabled."""
     if not config.enable_telemetry:
         return None
@@ -66,43 +58,3 @@ def _setup_telemetry(config: AgentConfig) -> t.Optional[trace_sdk.TracerProvider
     except Exception as e:
         logger.error(f"Failed to initialize telemetry: {e}")
         return None
-
-
-async def _create_agents(config: AgentConfig) -> AgentDict:
-    """Create and configure the ReAct agent."""
-    registry: AgentDict = {}
-
-    try:
-        logger.info("Creating all agents...")
-        registry["react"] = await create_react_agent(config)
-        registry["sql"] = await create_sql_agent(config)
-        return registry
-    except Exception as e:
-        logger.error(f"Failed to create agent: {e}")
-        raise AgentConfigError(f"Failed to create agent: {e}") from e
-
-class AgentRegistry:
-    """Registry of all agents."""
-    def __init__(
-        self,
-        config: AgentConfig,
-        agents: AgentDict = {},
-    ):
-        """Initialize registry."""
-        self.config = config
-        self.agents  = agents
-
-    @classmethod
-    async def create(cls, config: AgentConfig):
-        logger.info("Initializing the OSO agent registry...")
-        _setup_telemetry(config)
-        agents = await _create_agents(config)
-        registry = cls(config, agents)
-        logger.info("... agent registry ready")
-        return registry
-
-    def get_agent(self, name: str) -> BaseWorkflowAgent:
-        agent = self.agents.get(name)
-        if agent is None:
-            raise AgentMissingError(f"Agent '{name}' not found in the registry.")
-        return self.agents[name]
